# swagger.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  description: Swagger API for User Authentication API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /auth/register:
    post:
      summary: New user registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: craftlinkukraineteam_user
                password:
                  type: string
                  example: StrongPassword123
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
  /auth/login:
    post:
      summary: User Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: craftlinkukraineteam_user
                password:
                  type: string
                  example: StrongPassword123
      responses:
        200:
          description: Successful login - returns a JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
  /auth/profile:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: User Profile Data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
        401:
          description: Unauthorized access
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT